+++ docassemble_webapp/docassemble/webapp/file_access.py
@@ -33,7 +33,7 @@ def url_if_exists(file_reference, **kwargs):
     base_url = url_for('rootindex', _external=kwargs.get('_external', False)).rstrip('/')
     if len(parts) == 2:
         if cloud and docassemble.base.config.daconfig.get('use cloud urls', False):
-            m = re.search(r'^docassemble.playground([0-9]+)(.*)$', parts[0])
+            m = re.search(r'^docassemble\.playground([0-9]+)(.*)$', parts[0])
             if m:
                 user_id = m.group(1)
                 project = m.group(2)