+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [1.5.6] - 2024-09-01
+
+### Added
+- The `auto jinja filter` option under `features`.
+
+### Fixed
+- Adapted `syslog-ng` to Ubuntu 24.04.
+
 ## [1.5.5] - 2024-08-20
 
 ### Fixed